requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
dnspython
pydantic-settings
python-dotenv
google-generativeai
